Fix encoding when decoding inline encrypted data.
The problem with inline encryption is, that the encoding of the
decrypted data is not transfered. So we have to make assumtions. On the
one side gnupg tells, that it uses utf-8 by default and most of the
times this will be the correct encoding nowadays.
But you have the possibility to override the encoding for mime parts,
that should be followed, as otherwise the user has no possibilit to fix
broken encodings.